requirements.txt
schedule
requests
feedparser
pyyaml
discord.py